Use prop-types as peer-dependency. Don't transpile es6 version to es2015.
