+++ package.json
@@ -3,7 +3,7 @@
   "version": "3.0.0",
   "description": "Trigger state update on media query change.",
   "main": "lib/index.cjs.js",
-  "module": "lib/index.es.js",
+  "module": "src/index.js",
   "scripts": {
     "build": "rollup -c",
     "prepublish": "npm run build"
@@ -26,6 +26,7 @@
   },
   "homepage": "https://github.com/substantial/react-media-queryable",
   "peerDependencies": {
+    "prop-types": "^15.5.8",
     "react": "*"
   },
   "devDependencies": {
@@ -34,12 +35,7 @@
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "rollup": "^0.41.6",
-    "rollup-plugin-babel": "^2.7.1",
-    "rollup-plugin-commonjs": "^8.0.2",
-    "rollup-plugin-node-resolve": "^3.0.0"
-  },
-  "dependencies": {
-    "prop-types": "^15.5.8"
+    "rollup-plugin-babel": "^2.7.1"
   },
   "babel": {
     "presets": [
