+++ rollup.config.js
@@ -1,16 +1,11 @@
-import resolve from "rollup-plugin-node-resolve";
 import babel from "rollup-plugin-babel";
-import commonjs from "rollup-plugin-commonjs";
 
 export default {
   entry: "src/index.js",
   targets: [
-    { dest: "lib/index.cjs.js", format: "cjs" },
-    { dest: "lib/index.es.js", format: "es" }
+    { dest: "cjs/index.js", format: "cjs" }
   ],
   plugins: [
-    resolve(),
-    commonjs(),
     babel({
       exclude: "node_modules/**"
     })
